backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pypdf==5.0.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy==1.26.4
requests==2.32.3
orjson==3.10.7